`backend` -> `sync_cuda_with`, refine JAX example
